Collision object

+++ src/Types.hs
@@ -2,6 +2,6 @@ module Types where
 
 import           Linear (V4)
 
-type Color = V4 Int
+type Color = V4 Float
 type Point = V4 Float
 type Vector = V4 Float
